+++ src/app/(mobile-only)/login/page.tsx
@@ -1,11 +1,11 @@
 'use client';
 
-import LoginModal from '@/components/common/login-modal/login-modal';
+import LoginContent from '@/components/common/login-modal/login-content';
 import { useRouter } from 'next/navigation';
 
 export default function Page() {
 	const router = useRouter();
 	const previousPage = sessionStorage.getItem('previousPage');
 
-	return <LoginModal onClose={() => router.replace(previousPage)} />;
+	return <LoginContent onClose={() => router.replace(previousPage)} />;
 }
