+++ src/components/features/detail/content/detail-content.tsx
@@ -5,7 +5,7 @@ import { Suspense, useEffect, useState } from 'react';
 import DOMPurify from 'dompurify';
 import { getRelativeTime } from '@/lib/utils/getRelativeTime';
 import { categories } from '@/lib/constants/options';
-import LoginModal from '@/components/common/login-modal/login-modal';
+import LoginModal from '@/components/common/login-modal/login-content';
 import { useCurrentUserInfoStore } from '@/lib/store/useCurrentUserInfoStore';
 import parse, { Element } from 'html-react-parser';
 import { createNewsKick } from '@/services/apis/news/news.api';
