+++ src/components/layouts/root/navbar/login-portal.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import LoginModal from '@/components/common/login-modal/login-modal';
+import LoginModal from '@/components/common/login-modal/login-content';
 import { useIsLoginModalOpenStore } from '@/lib/store/useIsLoginModalOpenStore';
 import { useRouter } from 'next/navigation';
